pyproject.toml
[project]
name = "genai-template"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.89.0",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "langfuse>=4.0.6",
    "langgraph>=0.1.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
  "httpx",
  "mypy",
  "pre-commit",
  "pytest",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/genai_template"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
